tech stuff
git-svn-id: svn://192.168.1.10/public/CoffeeMud@16611 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/interfaces/TechLibrary.java
@@ -192,12 +192,12 @@ public interface TechLibrary extends CMLibrary
 	/**
 	 * If the two given objects are within an appropriate distance
 	 * from each other, this will return the correct amount of
-	 * accelleration force applied by the second object to the 
-	 * first.
+	 * accelleration g-force applied by the second object to the 
+	 * first. typically <= 1G
 	 * 
 	 * @param S the object being pulled
 	 * @param cO the object pulling
 	 * @return the amount of gravity force, or 0
 	 */
-	public long getGravityForce(SpaceObject S, SpaceObject cO);
+	public double getGravityForce(SpaceObject S, SpaceObject cO);
 }
